+++ test/css/test_css.rb
@@ -2,7 +2,7 @@
 
 require "helper"
 
-class TestNokogiriCSS < Nokogiri::TestCase
+describe Nokogiri::CSS do
   describe ".xpath_for" do
     it "accepts just a selector" do
       assert_equal(["//foo"], Nokogiri::CSS.xpath_for("foo"))
@@ -32,6 +32,24 @@ class TestNokogiriCSS < Nokogiri::TestCase
         ["./foo"],
         Nokogiri::CSS.xpath_for("foo", prefix: "./", visitor: Nokogiri::CSS::XPathVisitor.new),
       )
+    end
+
+    describe "error handling" do
+      it "raises a SyntaxError exception if the query is not valid CSS" do
+        assert_raises(Nokogiri::CSS::SyntaxError) { Nokogiri::CSS.xpath_for("'") }
+        assert_raises(Nokogiri::CSS::SyntaxError) { Nokogiri::CSS.xpath_for("a[x=]") }
+      end
+
+      it "raises a SyntaxError exception if the query is empty" do
+        e = assert_raises(Nokogiri::CSS::SyntaxError) { Nokogiri::CSS.xpath_for("") }
+        assert_includes("empty CSS selector", e.message)
+      end
+
+      it "raises an TypeError exception if the query is not a string" do
+        assert_raises(TypeError) { Nokogiri::CSS.xpath_for(nil) }
+        assert_raises(TypeError) { Nokogiri::CSS.xpath_for(3) }
+        assert_raises(TypeError) { Nokogiri::CSS.xpath_for(Object.new) }
+      end
     end
   end
 end